+++ python/acl_anthology/collections/event.py
@@ -90,6 +90,9 @@ class Talk:
 class Event:
     """An event, such as a meeting or a conference.
 
+    Note:
+        To manually create a new event, use [`Collection.create_event()`][acl_anthology.collections.collection.Collection.create_event].
+
     Attributes: Required Attributes:
         id: The ID of this event.
         parent: The Collection object that this event belongs to.