+++ python/CHANGELOG.md
@@ -6,7 +6,8 @@
 
 - Papers are now indexed by their bibkeys and can be retrieved via `Anthology.get_paper_by_bibkey()`.
 - Bibkeys can now be generated and updated, guaranteeing uniqueness.
-- Collections, Volumes, and Papers can now be newly created with functions on their respective parent objects.
+- Collections, Volumes, Papers, and Events can now be newly created with functions on their respective parent objects.
+  - Event creation currently has some unintuitive behaviour due to the existence of implicit event creation and linking; see docs.
 
 ### Changed